+++ CHANGELOG.md
@@ -4,9 +4,28 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [4.6.0] - 2022-05-15
+## [4.9.0] - 2022-08-14
+### Added
+- Moulinette Shortcuts (presets)
+
+## [4.9.0] - 2022-08-20
+### Added
+- Configuration to enable/disable available content from Moulinette Cloud
+
+## [4.8.0] - 2022-07-24
 ### Fixed
-- 4.5.1: #40 Non-host GMs can't use Moulinette for games hosted on The Forge
+- 4.8.1 : fix packaging
+- 4.8.2 : fix prefabs tab not working any more
+### Added
+- Moulinette API (searchUI & assetPicker)
+
+## [4.7.0] - 2022-06-10
+### Added
+- Support for thumb sizes
+
+## [4.6.1] - 2022-06-09
+### Fixed
+- 4.6.1: Fix confusing Moulinette message when supporting platinum creators
 ### Changed
 - Moulinette UI vertically resizable
 - Manage sources (for indexing process)